rendering node circles d3
