+++ app/stylesheets/routes.css
@@ -1,4 +1,4 @@
-.RouteCircle {
+/* .RouteCircle {
   font-family: Arial, Helvetica, sans-serif;
   color: white;
   width: 4em;
@@ -59,4 +59,4 @@
 #routeText:hover, #routeTimeDiff:hover {
   color: white;
   text-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
-}
+} */
